Refactor: Rename shimmers to placeholders and add list shimmer
This commit renames `shimmers.dart` to `placeholders.dart` and the `imageShimmerContainer` function to `imageShimmerPlaceholder` for clarity.

It also introduces a new `ListShimmerPlaceholder` widget which displays a shimmering list of placeholder boxes. This placeholder is now used on the post list page during initial loading.

Additionally, a commented-out line for retrieving the Firebase messaging token in `login.dart` has been uncommented.